Male izmene oko dodavanja prijatelja i ponistavanja zahteva za prijateljstvo.

+++ ISA/src/main/java/com/isa/service/FriendsServiceImpl.java
@@ -75,6 +75,7 @@ public class FriendsServiceImpl implements FriendsService{
 
 	@Override
 	public void declineFriendRequest(Long id) {
+		if(friendsRepository.findOne(id) == null) throw new IllegalArgumentException("Friendship doesn't exist");
 		friendsRepository.delete(id);
 	}
